+++ History.md
@@ -1,8 +1,16 @@
-0.9.2 / 2015-02-19
+0.9.4 / 2015-02-22
+==================
+- Fixed bug in rename method where renames of non-existing entities corrupted object.
+
+0.9.3 / 2015-02-22
 ==================
 - Merged mapValues and mapProperties methods into one map method with params and
   added an inplace flag
 - Added fromXml method
+
+0.9.2 / 2015-02-17
+==================
+- MapValues method
 
 0.9.1 / 2015-02-18
 ==================